renamed new_client -> client
